Moved GeometricUtils out of game.impl package, because they are general for game model
