+++ client/src/main/scala/it/cwmp/client/view/game/GameViewActor.scala
@@ -3,6 +3,7 @@ package it.cwmp.client.view.game
 import akka.actor.{Actor, ActorRef, Cancellable}
 import it.cwmp.client.controller.game.GameEngine
 import it.cwmp.client.model.DistributedState
+import it.cwmp.client.model.game.GeometricUtils
 import it.cwmp.client.model.game.impl._
 import it.cwmp.client.view.game.GameViewActor._
 import it.cwmp.client.view.game.model.{CellView, TentacleView}
